+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fixed task distribution counting infinite tasks and single mode tasks when calculating how many replicators can be on
   task
+* Infinite tasks are now shown as infinite in the task screen
 
 # 1.2.1